# More optimizations for InFlightRequest