--- a/src/Benchmarks/MicroBenchmarks/MicroBenchmarks.csproj
+++ b/src/Benchmarks/MicroBenchmarks/MicroBenchmarks.csproj
@@ -7,6 +7,7 @@
     <PlatformTarget>AnyCPU</PlatformTarget>
     <DebugType>pdbonly</DebugType>
     <DebugSymbols>true</DebugSymbols>
+    <IsPackable>false</IsPackable>
   </PropertyGroup>
   
   <PropertyGroup Condition="$(Configuration) == 'Release'">