--- a/src/Benchmarks/MicroBenchmarks/Program.cs
+++ b/src/Benchmarks/MicroBenchmarks/Program.cs
@@ -19,7 +19,7 @@ namespace MicroBenchmarks
     {
         public static void Main(string[] args)
         {
-            var summary = BenchmarkRunner.Run<NuidBenchmark>();
+            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
         }
     }
 }