--- a/src/NATS.Client/Conn.cs
+++ b/src/NATS.Client/Conn.cs
@@ -158,46 +158,6 @@ namespace NATS.Client
         private readonly ConcurrentDictionary<string, InFlightRequest> waitingRequests
             = new ConcurrentDictionary<string, InFlightRequest>(StringComparer.OrdinalIgnoreCase);
 
-        // Handles in-flight requests when using the new-style request/reply behavior
-        private sealed class InFlightRequest : IDisposable
-        {
-            internal InFlightRequest(string id, CancellationToken token, int timeout, Action<string> onCompleted)
-            {
-                this.Id = id;
-                this.Waiter = new TaskCompletionSource<Msg>();
-                this.onCompleted = onCompleted;
-                this.tokenSource = token == CancellationToken.None
-                    ? new CancellationTokenSource()
-                    : CancellationTokenSource.CreateLinkedTokenSource(token);
-                
-                this.tokenRegistration = this.tokenSource.Token.Register(() =>
-                {
-                    if (timeout > 0)
-                        this.Waiter.TrySetException(new NATSTimeoutException());
-
-                    this.Waiter.TrySetCanceled();
-                });
-
-                if(timeout > 0)
-                    this.tokenSource.CancelAfter(timeout);
-            }
-
-            public string Id { get; }
-            public TaskCompletionSource<Msg> Waiter { get; }
-            public CancellationToken Token => tokenSource.Token;
-
-            private readonly Action<string> onCompleted;
-            private readonly CancellationTokenSource tokenSource;
-            private CancellationTokenRegistration tokenRegistration;
-
-            public void Dispose()
-            {
-                this.tokenRegistration.Dispose();
-                this.tokenSource?.Dispose();
-                this.onCompleted?.Invoke(this.Id);
-            }
-        }
-
         // Prepare protocol messages for efficiency
         private byte[] PING_P_BYTES = null;
         private int    PING_P_BYTES_LEN;