… src/NATS.Client/Properties/AssemblyInfo.cs